Huh I guess I got carried away.
Refactor... everything really. The Flask server is broken.

Introduce "states" and "scenes". A state represents the current
configuration, and a scene is a "change" to that configuration.
All states are scenes, but a scene might be missing information
or include extra commands for the command generator to run.

I am also generating a list of commands for the wiim_controller to run.
This list will be executed in order, after validation. This is really
useful to test if the "diff" logic works.

Conceptually these are the steps to set a scene:
1. Get the current state (including airplay devices if relevant)
2. Apply the scene over the current state to get a destination state.
3. Use a command generator to create a list of command which get us from
   the current state to the destination state.
4. Run those commands on wiim_controller. This is what ends up making
   API requests.

There are a few other utilities which are included.
- A JSON "normalizer" (name TBD). This uses a JSON schema (or spec) to
  validate that an input dictionary is in the expected format. It will
  also apply some transformations to the dictionary, so that they are
  easier to write.

  I use the normalizer for states, scenes, and commands. As an example
  this input scene:
  ```
  { "output": "airplay" }
  ```
  will create this normalized scene (using default_value and
  default_key), which will enable all airplay devices:
  ```
  { "output": { "mode": "airplay", "airplay": [{"selected": true}] }
  ```

- A command runner. This takes a controller object, and a list of
  commands. It also has some builtins (print, sleep). It will execute
  the commands on the controller object, then try a buildin command.